.dagger/pyproject.toml
[project]
name = "codetanks-battlemaster"
version = "0.1.0"
requires-python = "~=3.13"
dependencies = [
    "dagger-io",
    "jinja2>=3.1.6",
    "toml>=0.10.2",
]

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }

[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"
